Create Interleaving String
First we can consider this as a recursive problem.
If  s3[i1+i2+1] == s1[i1+1], search(i1+1,i2)
If  s3[i1+i2+1] == s2[i2+1], search(i1,i2+1)
until end. However, this can only pass the small test cases, but failed on the large test cases.

So we need to think of the dynamic programming (DP), which usually have much less complexity. In this problem, a 2D DP is more suitable.  As usual, the typical way of solving dp is to find the state, and the optimal function. Here, the state can be considered as: A[i][j], which means S3[i+j-1] can be formed by S1[i-1] and S2[j-1] (for simplicity here string starts from 1, in the code we need to deal with that string starts from 0).

So, we have the optimal function:
  A[i][j] =   (s3[i+j-1]==s1[i-1]  && match[i-1][j])  || (s3[i+j-1] ==s2[j-1] && match[i][j-1])
also need to consider the A[i][0] and A[0][j] boundary condition.

Given the state and the optimal function, we also need to be careful with the initialization. Details see the code below.